w1retap-0.0.11.tar.gz
sha512sum c656192a5a3e832ab8b7e6a87bf8163f0e81b57aadeefc9ebf1344e38714e3c1a6ef41cfcebef1452535327bd1a19b1a220228d90dd3076d195558c277294d11

+++ src/libusblinux300/counter.c
@@ -52,7 +52,7 @@ int main(int argc, char **argv)
    int NumDevices=0;
    int i;
    int CounterPage;
-   ulong Count;
+   uint Count;
    int portnum=0;
 
    //------------------------------------------------------
