+++ src/libusblinux300/shacopr.c
@@ -35,7 +35,7 @@
 extern SMALLINT owFirst(int,SMALLINT,SMALLINT);
 extern SMALLINT owAcquire(int,char *);
 extern void owRelease(int);
-extern long msGettick();
+extern int msGettick();
 
 static void GetBytes(char* msg, uchar* buffer, int len, uchar defValue,
                      SMALLINT printHex);
@@ -52,7 +52,7 @@ static int GetSecret(char* name, uchar** secret);
 //
 int main(int argc, char** argv)
 {
-   long lvalue = 0;
+   int lvalue = 0;
    SHACopr copr;
    int namelen, auxlen;
    uchar coprFile[255];
@@ -302,7 +302,7 @@ int main(int argc, char** argv)
 static int GetSecret(char* name, uchar** secret)
 {
    uchar inputBuffer[255];
-   long lvalue=1, length;
+   int lvalue=1, length;
 
    printf("How would you like to enter the %s?\n", name);
    EnterNum("\n   1) Hex\n   2) Text\n", 1, &lvalue, 1, 2);
